requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml>=6.0.0
urllib3==2.2.1
python-dateutil==2.8.2
pandas>=2.0.0
openpyxl>=3.1.0
dateparser==1.2.1
numpy>=1.24.0
matplotlib>=3.7.0
glom==20.11.0
